refactor: 11.15에 설계된 refactoring 적용
