+++ client/src/components/common/MainSection.jsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import styled from 'styled-components';
-import MainHeader from '../SectionTitle';
+import MainHeader from './SectionTitle';
 import { BREAK_POINT_TABLET } from '../../magic-number';
 
-const MainSection = ({title, children}) => {
+const MainSection = ({ title, children }) => {
   return (
     <Main>
       <MainHeader title={title} />
