+++ client/src/components/make-section/make-section-components/EditorWithPreview.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import MdParser from '../../MdParser';
+import MdParser from '../../common/MdParser';
 import { flexBox } from '../../../styles/styled-components/mixin';
 
 const EditorWithPreview = ({ docData, docDispatch }) => {
