+++ client/src/components/make-section/make-section-components/Preview.jsx
@@ -1,20 +1,20 @@
-import React from "react";
+import React from 'react';
 import styled from 'styled-components';
-import MdParser from "../../MdParser";
+import MdParser from '../../common/MdParser';
 
-const Preview = ({docData}) => {
+const Preview = ({ docData }) => {
   return (
     <PreviewWrap>
       <MdParser content={docData.content} />
     </PreviewWrap>
-  )
-}
+  );
+};
 
 const PreviewWrap = styled.div`
   width: 100%;
   height: 432px;
   overflow: auto;
-  border: 1px solid #D7D7D7;
+  border: 1px solid #d7d7d7;
   box-sizing: border-box;
   border-radius: 10px;
 `;
