+++ client/src/reducer/doc-data-reducer.ts
@@ -57,7 +57,20 @@ export const docDataReducer = (state: DocData, action: any): DocData => {
     case 'INPUT_LINK':
       return { ...state, link: action.link };
     case 'INPUT_UPDATE_DATA':
-      return {...state, generation: action.generation, boostcamp_id: action.boostcamp_id, name: action.name, content: action.content, user_image: action.user_image, nickname: action.nickname, location: action.location, language: action.language, mbti: action.mbti, field: action.field, link: action.link };
+      return {
+        ...state,
+        generation: action.generation,
+        boostcamp_id: action.boostcamp_id,
+        name: action.name,
+        content: action.content,
+        user_image: action.user_image,
+        nickname: action.nickname,
+        location: action.location,
+        language: action.language,
+        mbti: action.mbti,
+        field: action.field,
+        link: action.link,
+      };
     default:
       return state;
   }
